RestaurantPage에 container를 도입하라
