+++ src/pages/RestaurantPage.jsx
@@ -1,23 +1,4 @@
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { useParams } from 'react-router-dom';
-
-import { loadRestaurantInfo } from '../redux_module/asyncActions';
-
-export default function RestaurantPage() {
-  const dispatch = useDispatch();
-  const { id: restaurantID } = useParams();
-
-  useEffect(() => {
-    dispatch(loadRestaurantInfo(restaurantID));
-  }, []);
-
-  const { name, address, menuItems } = useSelector(
-    (state) => state.restaurant.information,
-  );
-
-  // Todo container/presentational 도입(?)
-
+export default function RestaurantPage({ name, address, menuItems }) {
   return (
     <>
       <h2>{name}</h2>
